First commit because git didn't work but it doesn't change much. Documentations with Examples and pdf and App Separations from Documentations in TESTS are not included in git.
In FirstTry is the first try with the MMApp but after finding the Examples in Documentations I got the idea how to create an KIVY App correctly so the directory App has the right App. Now it has a functioning register_screen and two tabs with Transition between them. It took a lot of tries ad fails but the structure is now done and is working. It's still a skeleton, but a correct one.